create args and data modules
