+++ src/cnlpt/api/utils.py
@@ -19,7 +19,7 @@ from transformers import (
 )
 from transformers.tokenization_utils import PreTrainedTokenizer
 
-from ..cnlp_data import cnlp_preprocess_data
+from ..data.preprocess import cnlp_preprocess_data
 from ..models import CnlpConfig
