+++ src/cnlpt/data/tasks.py
@@ -1,0 +1,4 @@
+mtl = "mtl"
+classification = "classification"
+tagging = "tagging"
+relex = "relations"
